+++ api.go
@@ -6,43 +6,36 @@ import (
 	"time"
 )
 
-type CodecType interface {
-	NewEncoder(w io.Writer) Encoder
-	NewDecoder(r io.Reader) Decoder
+type Protocol interface {
+	NewCodec(rw io.ReadWriter) Codec
 }
 
-type Encoder interface {
-	Encode(msg interface{}) error
+type Codec interface {
+	Receive() (interface{}, error)
+	Send(interface{}) error
+	Close() error
 }
 
-type Decoder interface {
-	Decode(msg interface{}) error
-}
-
-type Disposeable interface {
-	Dispose()
-}
-
-func Serve(network, address string, codecType CodecType) (*Server, error) {
+func Serve(network, address string, protocol Protocol, sendChanSize int) (*Server, error) {
 	listener, err := net.Listen(network, address)
 	if err != nil {
 		return nil, err
 	}
-	return NewServer(listener, codecType), nil
+	return NewServer(listener, protocol, sendChanSize), nil
 }
 
-func Connect(network, address string, codecType CodecType) (*Session, error) {
+func Connect(network, address string, protocol Protocol, sendChanSize int) (*Session, error) {
 	conn, err := net.Dial(network, address)
 	if err != nil {
 		return nil, err
 	}
-	return NewSession(conn, codecType), nil
+	return NewSession(protocol.NewCodec(conn), sendChanSize), nil
 }
 
-func ConnectTimeout(network, address string, timeout time.Duration, codecType CodecType) (*Session, error) {
+func ConnectTimeout(network, address string, timeout time.Duration, protocol Protocol, sendChanSize int) (*Session, error) {
 	conn, err := net.DialTimeout(network, address, timeout)
 	if err != nil {
 		return nil, err
 	}
-	return NewSession(conn, codecType), nil
+	return NewSession(protocol.NewCodec(conn), sendChanSize), nil
 }