+++ example/json_toy/main.go
@@ -1,0 +1,71 @@
+package main
+
+import (
+	"log"
+
+	"github.com/funny/link"
+	"github.com/funny/link/codec"
+)
+
+type AddReq struct {
+	A, B int
+}
+
+type AddRsp struct {
+	C int
+}
+
+func main() {
+	json := codec.Json()
+	json.Register(AddReq{})
+	json.Register(AddRsp{})
+
+	server, err := link.Serve("tcp", "0.0.0.0:0", json, 0 /* sync send */)
+	checkErr(err)
+	go serverLoop(server)
+	addr := server.Listener().Addr().String()
+
+	client, err := link.Connect("tcp", addr, json, 0)
+	checkErr(err)
+	clientLoop(client)
+}
+
+func serverLoop(server *link.Server) {
+	for {
+		session, err := server.Accept()
+		checkErr(err)
+		go sessionLoop(session)
+	}
+}
+
+func sessionLoop(session *link.Session) {
+	for {
+		req, err := session.Receive()
+		checkErr(err)
+
+		err = session.Send(&AddRsp{
+			req.(*AddReq).A + req.(*AddReq).B,
+		})
+		checkErr(err)
+	}
+}
+
+func clientLoop(session *link.Session) {
+	for i := 0; i < 10; i++ {
+		err := client.Send(&AddReq{
+			i, i,
+		})
+		checkErr(err)
+		log.Printf("Send: %d + %d", i, i)
+
+		rsp, err := client.Receive()
+		checkErr(err)
+		log.Printf("Receive: %d", rsp.(*AddRsp).C)
+	}
+}
+
+func checkErr(err error) {
+	if err != nil {
+		log.Fatal(err)
+	}
+}